added saving to database and greeting page (todo: other crud operations + pagination)

+++ src/main/java/ru/jrintership/TestJRInternship/services/Impl/UserServiceImpl.java
@@ -41,12 +41,12 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User findById(Long id) {
+    public User findById(Integer id) {
         return userRepository.getOne(id);
     }
 
     @Override
-    public void deleteUser(Long id) {
+    public void deleteUser(Integer id) {
         userRepository.deleteById(id);
     }
